Flattened linode_api.objects.networking into a single file
