--- linode_api/objects/networking/__init__.py
@@ -1,3 +1,0 @@
-from .ipaddress import IPAddress
-from .ip6pool import IPv6Pool
-from .ip6range import IPv6Range
