--- linode_api/objects/networking/ip6range.py
@@ -1,13 +1,0 @@
-from __future__ import absolute_import
-
-from linode_api.objects import Base, Property, Region
-
-
-class IPv6Range(Base):
-    api_endpoint = '/networking/ipv6/ranges/{}'
-    id_attribute = 'range'
-
-    properties = {
-        'range': Property(identifier=True),
-        'region': Property(slug_relationship=Region, filterable=True),
-    }
